+++ src/epics/index.js
@@ -9,10 +9,15 @@ import {
   getDocContentEpic,
 } from './docs';
 
+import {
+  getPackageLatestVersion,
+} from './alert';
+
 export const rootEpic = combineEpics(
   getComponentsEpic,
   getMarkupEpic,
   getDocContentEpic,
+  getPackageLatestVersion,
 );
 
 export default rootEpic;