+++ src/reducers/atomic.js
@@ -1,3 +1,5 @@
+import p from 'immer';
+
 import {
   SET_COMPONENTS,
   SET_COMPONENT_MARKUP,
@@ -6,50 +8,27 @@ import {
 
 const defaultState = {};
 
-export default function atomicReducer(state = defaultState, action) {
+const atomicReducer = p((state = defaultState, action) => {
   switch (action.type) {
 
     // Add the base components collection
     case SET_COMPONENTS:
-      return {
-        ...state,
-        sources: action.payload,
-      };
+      state.sources = action.payload;
+      state.types = Object.keys(action.payload);
+      break;
 
     // Update a specific component
     case SET_COMPONENT_MARKUP:
-      return {
-        ...state,
-        sources: {
-          ...state.sources,
-          [action.payload.type]: [
-            ...state.sources[action.payload.type].slice(0, action.payload.id),
-            action.payload,
-            ...state.sources[action.payload.type].slice(action.payload.id + 1),
-          ],
-        },
-      };
+      state.sources[action.payload.type][action.payload.id].content = action.payload.content;
+      break;
 
     // Update a specific variant inside it parent component
     case SET_VARIANT_MARKUP:
-      return {
-        ...state,
-        sources: {
-          ...state.sources,
-          [action.payload.type]: [
-            ...state.sources[action.payload.type].slice(0, action.payload.parent_id),
-            {
-              ...state.sources[action.payload.type][action.payload.parent_id],
-              variants: [
-                ...state.sources[action.payload.type][action.payload.parent_id].variants.slice(0, action.payload.id),
-                action.payload,
-                ...state.sources[action.payload.type][action.payload.parent_id].variants.slice(action.payload.id + 1),
-              ],
-            },
-            ...state.sources[action.payload.type].slice(action.payload.parent_id + 1),
-          ],
-        },
-      };
+      state.sources[action.payload.type][action.payload.parent_id].variants[[action.payload.id]].content = action.payload.content;
+      break;
+
     default: return state;
   }
-}
+});
+
+export default atomicReducer;